LPD-55424 Reorganize FDS tests
