+++ modules/test/playwright/tests/frontend-data-set-web/main/tests/external/commerce.spec.ts
@@ -5,12 +5,12 @@
 
 import {expect, mergeTests} from '@playwright/test';
 
-import {apiHelpersTest} from '../../../fixtures/apiHelpersTest';
-import {applicationsMenuPageTest} from '../../../fixtures/applicationsMenuPageTest';
-import {commercePagesTest} from '../../../fixtures/commercePagesTest';
-import {dataApiHelpersTest} from '../../../fixtures/dataApiHelpersTest';
-import {loginTest} from '../../../fixtures/loginTest';
-import getRandomString from '../../../utils/getRandomString';
+import {apiHelpersTest} from '../../../../../fixtures/apiHelpersTest';
+import {applicationsMenuPageTest} from '../../../../../fixtures/applicationsMenuPageTest';
+import {commercePagesTest} from '../../../../../fixtures/commercePagesTest';
+import {dataApiHelpersTest} from '../../../../../fixtures/dataApiHelpersTest';
+import {loginTest} from '../../../../../fixtures/loginTest';
+import getRandomString from '../../../../../utils/getRandomString';
 
 const test = mergeTests(
 	apiHelpersTest,
